% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{acf_plot}
\alias{acf_plot}
\title{Residual autocorrelation plot}
\usage{
acf_plot(baseline_obj, lag_max = 12)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline model function}

\item{lag_max}{The maximum lag at which to calculate the acf (autocorrelation function)}
}
\value{
A ggplot object
}
\description{
\code{acf_plot} A function that create an autocorrelation plot of the baseline model residual
}
