% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{act_plot}
\alias{act_plot}
\title{Plot of the actual pre or post installation data}
\usage{
act_plot(Data, title = NULL, low_range_0 = NULL, high_range_0 = NULL,
  low_range_T_0 = NULL, high_range_T_0 = NULL)
}
\arguments{
\item{Data}{A data frame of the actual observations}

\item{low_range_0}{a numerical value of the lower bound of the eload axis}

\item{high_range_0}{a numerical value of the upper bound of the eload axis}

\item{low_range_T_0}{a numerical value of the lower bound of the Temp axis}

\item{high_range_T_0}{a numerical value of the upper bound of the Temp axis}
}
\value{
A dygraph time series plot object
}
\description{
\code{act_plot} a function that generates a dygraph plot of the actual observations (eload and Temp)
}
