% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{act_vs_fit_plot}
\alias{act_vs_fit_plot}
\title{actual vs. fitted scatter plot}
\usage{
act_vs_fit_plot(baseline_obj)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline model function}
}
\value{
A ggplot object
}
\description{
\code{act_vs_fit_plot} A function that create a scatterplot of the actual vs. fitted data produced by the baseline model
}
