% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{axis_range}
\alias{axis_range}
\title{Compute plot axis range for two different data files (e.g, pre and post)}
\usage{
axis_range(Data_1, Data_2)
}
\arguments{
\item{Data_1}{and Data_2 dataframes of two different data files}
}
\value{
a list object with the following components:
\describe{
  \item{low_range_0}{a numerical value of the lower bound of the eload axis}
  \item{high_range_0}{a numerical value of the upper bound of the eload axis}
  \item{low_range_T_0}{a numerical value of the lower bound of the Temp axis}
  \item{high_range_T_0}{a numerical value of the upper bound of the Temp axis}
}
}
\description{
\code{axis_range} Compute plot axis range for two different data files (e.g, pre and post)
}
