% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{clean_Temp_2}
\alias{clean_Temp_2}
\title{Clean the Temperature data}
\usage{
clean_Temp_2(Data, thresh = 5)
}
\arguments{
\item{thresh}{A numeric that correspond to the number of similar successive
temperature values}

\item{data}{A dataframe of training or prediction data.}
}
\value{
A dataframe that correspond to the cleaned training or prediction data
}
\description{
\code{clean_Temp_2} This function remove the period of time where temperature
observations do not change.
}
