% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{convert_15min_to_1_hour}
\alias{convert_15min_to_1_hour}
\title{Convert 15 minute interval data into hourly interval data}
\usage{
convert_15min_to_1_hour(Data, kWh = TRUE)
}
\arguments{
\item{Data}{A dataframe that contains 15 minutes interval data}

\item{kWh}{A logical if TRUE it means that the interval data are energy data}
}
\value{
A dataframe with hourly interval data
}
\description{
\code{convert_15min_to_1_hour} This function convert 15 minute interval data into hourly interval data
}
