% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrevent.R
\name{cpt_det}
\alias{cpt_det}
\title{Function to detect potential non-routine events}
\usage{
cpt_det(baseline_obj, interval)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline
function (e.g., towt_baseline, gbm_baseline)}
}
\value{
a cpt_det object, which is a list with the following components:
\describe{
  \item{savings}{vector corresponding to the estimated savings time series}
  \item{nre}{Boolean indicating if non-routine events were identified}
  \item{cpts}{changepoints indices}
  \item{interval_means}{mean values of detected intervals}
  \item{interval_vars}{variance values of detected intervals}
}
}
\description{
\code{cpt_det} This function identifies potential non-routine events in the savings
time series it uses cpt.meanvar function from changepoint R package to identify
changes in mean and variance
}
