% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{cpt_save_plot}
\alias{cpt_save_plot}
\title{Plot the residual/savings and teh idetified change points in the post period}
\usage{
cpt_save_plot(baseline_obj, cpt_det_obj, title = NULL, plot_group = NULL)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline function}

\item{title}{a character string that corresponds to the main title of the plot}
}
\value{
A dygraph time series plot
}
\description{
\code{cpt_save_plot} Read a baseline object and plots the residuals between the actual and the predicted data
}
