% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{create_date_var}
\alias{create_date_var}
\title{Create an input variable corresponding to given intervals}
\usage{
create_date_var(Data, intervals_path = NULL, start = NULL, end = NULL,
  var_name = "date_off")
}
\arguments{
\item{Data}{A dataframe of pre or post-installation data.}

\item{intervals_path}{A path to the file with intervals to extract}

\item{start}{A vector of start date of each interval}

\item{end}{A vector of end date of each interval. Note that start vector and
end vector should have the same length}

\item{var_name}{A string character that correspond to the name of the new
input variable}
}
\value{
a dataframe with the additional input variable
}
\description{
\code{create_date_var} This function Create a binary input variable corresponding
to given intervals if the time stamps correspond to a date within the given
interval then the value of the input variables will be equal to 1 and if not it
will be equal to 0
}
