% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_load}
\alias{data_load}
\title{Load data into the shiny application}
\usage{
data_load(files_path, files_names, var_out, Post = T, clean = T)
}
\arguments{
\item{files_path}{a list that contains all the paths of the considered data files}

\item{files_names}{a list that contains all the files names}

\item{var_out}{a shiny reactiveValues object where the data are stored}

\item{Post}{Boolean that determine if the data is post-installation or
post-installation. If true then the data is considered as post-installation.}

\item{clean}{Boolean that determine if an automatic data cleaning is performed.}
}
\value{
var_out reactiveValues object with a new object where the loaded data are stored
}
\description{
\code{data_load} This function is used by the shiny application to load
pre or post dataset and store it within a shiny reactiveValues object
}
