% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{eload_heatmap}
\alias{eload_heatmap}
\title{Eload heatmap}
\usage{
eload_heatmap(Data, title = NULL, colors = colorRamps::matlab.like(100),
  zauto = T, zmin = 0, zmax = 0)
}
\arguments{
\item{Data}{a data frame to plot}
}
\value{
A plot_ly object
}
\description{
\code{eload_heatmap} A function that create a heatmap plot of eload where xaxis corresponds to the time of the week and yaxis corresponds to the weeks
}
