% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{eload_vs_input_plot}
\alias{eload_vs_input_plot}
\title{eload vs. Temperature scatter plot}
\usage{
eload_vs_input_plot(Data, input = "Temp")
}
\arguments{
\item{Data}{a dataframes with eload and Temp columns}
}
\value{
A ggplot object
}
\description{
\code{eload_vs_Temp_plot} A function that creates a scatterplot of eload vs. Temperature
}
