% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{errors_vs_input_plot}
\alias{errors_vs_input_plot}
\title{errors vs. Temperature scatter plot}
\usage{
errors_vs_input_plot(baseline_obj, input = "Temp")
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline model function}
}
\value{
A ggplot object
}
\description{
\code{errors_vs_Temp_plot} A function that creates a scatterplot of the baseline model errors vs. Temperature  data produced by the baseline model
}
