% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{fsu_estimation_sc}
\alias{fsu_estimation_sc}
\title{Potential fractional savings}
\usage{
fsu_estimation_sc(screen_out, d_post, frac_savings)
}
\arguments{
\item{screen_out}{a shiny reactiveValues object where the baseline object and the pre data are stored}

\item{d_post}{a numerical value corresponding to the number of required days in the post period}

\item{frac_savings}{a numerical value corresponding to the user defined fractional savings}
}
\value{
a dataframe of the estimated fractional savings uncertainties at three different confidence levels
}
\description{
\code{savings_summary} This function is used by the shiny application
to produce an estimation of potential fractional savings uncertainties when the screening analysis is performed
}
