% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{k_dblocks_cv}
\alias{k_dblocks_cv}
\title{K-fold-day cross validation function.}
\usage{
k_dblocks_cv(Data, k_folds)
}
\arguments{
\item{Data}{A dataframe.}

\item{k_folds}{An integer that corresponds to the number of CV folds.}
}
\value{
A list of row indexes corresponding to the training data.
}
\description{
\code{k_dblocks_cv} splits the data into k folds by randomly selecting blocks of data,
where each block correspond to a calendar day.
}
