% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_baseline.R
\name{k_wblocks_cv}
\alias{k_wblocks_cv}
\title{K-fold-week cross validation function.}
\usage{
k_wblocks_cv(Data, k_folds)
}
\arguments{
\item{Data}{A dataframe.}

\item{k_folds}{An integer that corresponds to the number of CV folds.}
}
\value{
A list of row indexes corresponding to the training data.
}
\description{
\code{k_wblocks_cv} splits the data into k folds by randomly selecting blocks of data,
where each block correspond to a calendar week.
}
