% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_session}
\alias{load_session}
\title{load a session}
\usage{
load_session(load_path)
}
\arguments{
\item{load_path}{the path of a saved session}
}
\description{
\code{load_session} This function is used by the shiny application to load
a saved session
}
