% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_of_days}
\alias{number_of_days}
\title{Compute number of days in the data}
\usage{
number_of_days(Data)
}
\arguments{
\item{A}{numeric that corresponds to the number of days}
}
\description{
\code{number_of_days} This function computes the number of days for which observations
are available
}
