% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{portfolio_savings}
\alias{portfolio_savings}
\title{Savings summary}
\usage{
portfolio_savings(sav_out, screened = F)
}
\arguments{
\item{sav_out}{a shiny reactiveValues object where the baseline object and the pre/post data are stored}

\item{screened}{a Boolean, which defines if the all the data are included in the portfolio analysis (screened = FALSE) or if only the data that have passed the screeneing are used (screened = TRUE)}
}
\value{
a portfolio_savings object which is a list  with the following components:
\describe{
  \item{Savings_portfolio}{numerical value corresponding to the estimated savings at portfolio level}
  \item{Delta_portfolio}{numerical value corresponding to the estimated savings uncertainty at portfolio level}
  \item{Savings_Range}{a text string corresponding to the estimated savings range at the portfolio level}
  \item{FS_portfolio}{numerical value corresponding to the estimated fractional savings at portfolio level}
  \item{FSU_portfolio}{numerical value corresponding to the estimated fractional savings uncertainty at portfolio level}
  \item{FS_Range}{a text string corresponding to the estimated fractional savings range at the portfolio level}
}
}
\description{
\code{savings_summary} This function is used by the shiny application
to produce results summary at the portfolio level
}
