% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{post_plot}
\alias{post_plot}
\title{Plot of the post-installation period data}
\usage{
post_plot(baseline_obj, title = NULL, plot_group = NULL,
  low_range_0 = NULL, high_range_0 = NULL, low_range_T_0 = NULL,
  high_range_T_0 = NULL)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline model function}

\item{low_range_0}{a numerical value of the lower bound of the eload axis}

\item{high_range_0}{a numerical value of the upper bound of the eload axis}

\item{low_range_T_0}{a numerical value of the lower bound of the Temp axis}

\item{high_range_T_0}{a numerical value of the upper bound of the Temp axis}
}
\value{
A dygraph time series plot object
}
\description{
\code{pre_plot} Read a baseline object and plots the actual and the predicted data
}
