% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pred_accuracy}
\alias{pred_accuracy}
\title{Prediction accuracy metrics computation}
\usage{
pred_accuracy(baseline_obj)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the towt_baseline or by gbm_baseline function}
}
\value{
A dataframe with where the columns correspond to R2, CV(RMSE) and NMBE
}
\description{
\code{pred_accuracy} This function compute the following prediction accuracy metrics:  R2, CV(RMSE) and NBME
}
