% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratio_missing}
\alias{ratio_missing}
\title{Estimate the ration of missing data}
\usage{
ratio_missing(Data, interval)
}
\arguments{
\item{Data}{A data frame that contains time column}
}
\value{
A numeric that corresponds to the missing values ratio
}
\description{
\code{ratio_missing} This function estimates the ratio of missing observations
by calculating the missing time steps
}
