% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_predictions}
\alias{save_predictions}
\title{Predictions Data Savings}
\usage{
save_predictions(model_obj_list, save_results_path = NULL, post = TRUE)
}
\arguments{
\item{model_obj_list}{A list of baseline objects}

\item{save_results_path}{The path where the files will be saved}

\item{post}{A boolean that determines if post prediction should be saved}
}
\description{
\code{save_predictions} This function extract and save predictions 
from baseline model objects
}
