% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_session}
\alias{save_session}
\title{Save a session}
\usage{
save_session(save_dir, project_name, results)
}
\arguments{
\item{save_dir}{the path of the folder where the session will be saved}

\item{project_name}{the name of the project (given by teh user)}

\item{the}{shiny reactiveValues object where the results are stored}
}
\description{
\code{save_session} This function is used by the shiny application to save the
 results of the session
}
