% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{savings}
\alias{savings}
\title{Function to calculate the savings of the post period.}
\usage{
savings(baseline_obj)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline function}
}
\value{
a savings object, which is a list with the following components:
\describe{
  \item{savings}{numeric corresponding to the estimated savings}
  \item{frac_savings}{numeric corresponding to the fractional savings}
}
}
\description{
\code{savings} This function computes the estimated savings using the predicted data of the post period.
}
