% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{savings_heatmap}
\alias{savings_heatmap}
\title{Savings heatmap}
\usage{
savings_heatmap(baseline_obj, title = NULL,
  colors = colorRamps::matlab.like(100), zauto = T, zmin = 0, zmax = 0)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline function}
}
\value{
A plot_ly object
}
\description{
\code{savings_heatmap} A function that creates a heatmap plot of savings where xaxis corresponds to the time of the week and yaxis corresponds to the weeks
}
