% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{savings_results_plot}
\alias{savings_results_plot}
\title{Savings barplot}
\usage{
savings_results_plot(savings_res_tab)
}
\arguments{
\item{savings_res_tab}{a data frame produced by the savings_results function}
}
\value{
A plot_ly object
}
\description{
\code{savings_results_plot} A function that creates a barplot of savings estimation and their uncertainties
}
