% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{savings_summary}
\alias{savings_summary}
\title{Savings summary}
\usage{
savings_summary(sav_out)
}
\arguments{
\item{sav_out}{a shiny reactiveValues object the baseline object and the pre/post data are stored}

\item{inCL}{a numerical value corresponding to the user specified confidence level}
}
\value{
a dataframe of the savings summary
}
\description{
\code{savings_summary} This function is used by the shiny application
to produce the savings and the uncertainties summary table
}
