% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings.R
\name{savings_uncert}
\alias{savings_uncert}
\title{Function to estimate the savings uncertainties using ASHRAE 14 definition}
\usage{
savings_uncert(baseline_obj, savings_obj, CL)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline function (e.g., towt_baseline)}

\item{savings_obj}{savings_obj object produced by the savings function}

\item{CL}{a numeric value that defines the confidence level at which the uncertainty is estimated}
}
\value{
a savings_uncert object, which is a list with the following components:
\describe{
  \item{delta_sav}{numeric corresponding to the estimated savings uncertainty}
  \item{fsu}{numeric corresponding to the estimated fractional savings uncertainty}
  \item{rho}{the estimated autocorrelation}
}
}
\description{
\code{savings_uncert} This function computes the estimated savings uncertainties using ASHRAE 14 definition
}
