% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{screen_pie_plot}
\alias{screen_pie_plot}
\title{Screening pie plot}
\usage{
screen_pie_plot(screen_summary_list)
}
\arguments{
\item{screen_summary_list}{a list object produced by screen_summary function}
}
\value{
A plot_ly object
}
\description{
\code{screen_pie_plot} A function that create a pie plot the screening results
}
