% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{screen_summary}
\alias{screen_summary}
\title{Screening summary}
\usage{
screen_summary(summary_tab, R2_trsh, CVRMSE_trsh, NMBE_trsh)
}
\arguments{
\item{summary_tab}{a dataframe produced by \code{train_model_summary} function}

\item{R2_trsh}{a numerical value corresponding to the R2 threshold}

\item{CVRMSE_trsh}{a numerical value corresponding to the CV(RMSE) threshold}

\item{NMBE_trsh}{a numerical value corresponding to the NMBE threshold}
}
\value{
a screen_summary object, which is a list with the following components:
\describe{
  \item{win_tab}{a summary_tab dataframe containing only the results for the data files that passed the screening}
  \item{los_tab}{a summary_tab dataframe containing only the results for the data files that did not passe the screening}
  \item{win_ratio}{a numerical value corresponding to the ratio of data files that passed the screening}
  \item{los_ratio}{a numerical value corresponding to the ratio of data files that did not passe the screening}
}
}
\description{
\code{screen_summary} This function is used by the shiny application
to perform a screening of the baseline models results
}
