% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_features}
\alias{time_features}
\title{Extract features from the time}
\usage{
time_features(data)
}
\arguments{
\item{Data}{A data frame that contains time column}
}
\value{
Data data frame that contains new columns corresponding to the extract time features:
\describe{
  \item{dts}{POSIXct object of the time stamps}
  \item{month}{the month of the time stamps}
  \item{wday}{the day of the week of the time stamps}
  \item{hour}{the hour of the time stamps}
  \item{minute}{the minute  of the time stamps}
  \item{tod}{the time of the day of the time stamps}
  \item{tow}{the time of the week of the time stamps}
  \item{date}{the date of the time stamps}
  \item{week}{the week of the year of the time stamps}
  \item{week_date}{the date of the sunday of the time stamps week}
}
}
\description{
\code{number_of_days} This function computes the number of days for which data
is available
}
