% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{time_format}
\alias{time_format}
\title{Convert the timestamps into the default format}
\usage{
time_format(Data, format)
}
\arguments{
\item{format}{A character string that defines the actual format of the timestamps.
Use the description of the base R function \emph{strptime} to define the format.}

\item{data}{A data frame that contains time column}
}
\value{
A data frame with timestamps converted into "%m/%d/%y %H:%M" format
}
\description{
\code{time_format} This function convert the actual timestamps format into "%m/%d/%y %H:%M" format
}
