% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{to_exclude}
\alias{to_exclude}
\title{Exclude data for given intervals}
\usage{
to_exclude(Data, intervals_to_exclude)
}
\arguments{
\item{Data}{A dataframe}

\item{intervals_to_exclude}{a path to the file with intervals to exclude}
}
\value{
A dataframe containing the remaining observations
}
\description{
\code{to_exclude} This function exclude data for given intervals
}
