% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{to_extract}
\alias{to_extract}
\title{Extract data from given intervals}
\usage{
to_extract(Data, intervals_to_extract = NULL, start = NULL, end = NULL)
}
\arguments{
\item{Data}{A dataframe containing all observations}

\item{intervals_to_extract}{A path to the file with intervals to extract}

\item{start}{The start date of the interval to exclude}

\item{end}{The end date of the interval to exclude}
}
\value{
a to_extract object, which is a list with the following components:
\describe{
  \item{Data}{a dataframe containing the remaining observations}
  \item{Data_new}{a dataframe containing the extracted observations}
}
}
\description{
\code{to_extract} This function extract data from given intervals
}
