% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towt_baseline.R
\name{towt_baseline}
\alias{towt_baseline}
\title{Time Of the Week and Temperature baseline model function.}
\usage{
towt_baseline(train_path = NULL, pred_path = NULL, train_Data = NULL,
  pred_Data = NULL, timescaleDays = 14, verbosity = 1,
  intervalMinutes = 15, fahrenheit = T)
}
\arguments{
\item{train_path}{The path of the file from which the training data are to be read from.}

\item{pred_path}{The path of the file from which the prediction data are to be read from.}

\item{train_Data}{A dataframe, of the training period, where the columns correspond to the time steps (time), the energy load (eload) and to the Temperature (Temp).}

\item{pred_Data}{A dataframe, of the prediction period, where the columns correspond to the time steps (time), the energy load (eload) and to the Temperature (Temp).}

\item{timescaleDays}{Numeric correspond to the timescale for weighting function.}

\item{verbosity}{Numeric that determine what progress and error reports to print .}

\item{intervalMinutes}{Numeric of length of a Time Of Week interval as input variables.}

\item{fahrenheit}{Boolean that determine if the temperatures are in  Fahrenheit.}
}
\value{
a towt_baseline object, which alist with the following components:
\describe{
  \item{towt_model}{an object that has been created by the function makeBaseline,
   and which correspond to the towt model.}
  \item{train}{a data frame that correspond to the training data after the
  cleaning and filtering function were applied.}
  \item{fitting}{the fitted values.}
  \item{goodness_of_fit}{a data frame that contains the goodness of fitting metrics.}
  \item{pred}{a data frame that correspond to the prediction data after the
  cleaning and filtering function were applied.}
  \item{prediction}{the predicted values.}
}
}
\description{
\code{towt_baseline} This function build a baseline model using gradient boosting machine algorithm.
}
