% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towt_baseline.R
\name{towt_time_var}
\alias{towt_time_var}
\title{Convert the time format}
\usage{
towt_time_var(Data)
}
\arguments{
\item{Data}{A dataframe of the training or prediction data.}
}
\value{
A dataframe where the time column is in the POSIXct format.
}
\description{
\code{towt_time_var} Function that exclude observation where the time stamps are missing and convert the time column into the POSIXct format.
}
