% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_model}
\alias{train_model}
\title{Train Baseline models}
\usage{
train_model(var_out, screen = T, Model = "TOWT", pam_list = NULL,
  days_off_path = NULL)
}
\arguments{
\item{var_out}{a shiny reactiveValues object where the taining data are stored}

\item{screen}{Boolean that determines if the analysis is a screening or a
savings analysis. If it's a screening then the baseline model will return only
the fitting results of the pre-installation. While if it's savings analysis the
baseline model will also return the prediction for the post-installation period}

\item{Model}{Character string that correspond to the name of the model}

\item{pam_list}{A list with information about the hyperparameters of the considel model}

\item{days_off_path}{A path of the file that include the date of the days off}
}
\value{
train_model a list
}
\description{
\code{train_model} This function is used by the shiny application to train
the baseline models for all the data that are stored in the var_out object
}
