% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{train_model_summary}
\alias{train_model_summary}
\title{Baseline models results summary}
\usage{
train_model_summary(model_obj_list, files_names)
}
\arguments{
\item{model_obj_list}{a list of the baselne models results for all the data}

\item{files_names}{a list of names of all the data files for which a baseline models have been built}
}
\value{
A data frame containing the results summary
}
\description{
\code{train_model_summary} This function is used by the shiny application
to summarise the baseline models results
}
