% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{weekly_save_barplot}
\alias{weekly_save_barplot}
\title{BarPlot the weekly savings}
\usage{
weekly_save_barplot(baseline_obj, title = NULL, plot_group = NULL)
}
\arguments{
\item{baseline_obj}{A baseline object, which is produced by the baseline function}

\item{title}{a character string that corresponds to the main title of the plot}
}
\value{
A dygraph time series plot
}
\description{
\code{weekly_save_barplot} Read a baseline object and plots the weekly residuals between the actual and the predicted data
}
