% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_interval}
\alias{detect_interval}
\title{Compute the granularity of the time series}
\usage{
detect_interval(Data)
}
\arguments{
\item{Data}{A data frame that contains time column}
}
\value{
A numeric that corresponds to the interval between the observations in minutes
}
\description{
\code{detect_interval} This function computes the granularity of the time series
}
