% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrevent.R
\name{nre_eval}
\alias{nre_eval}
\title{Function to identify change points in the whole dataset}
\usage{
nre_eval(var_out)
}
\arguments{
\item{var_out}{a shiny reactiveValues object where the data are stored}
}
\value{
a list
}
\description{
\code{train_model} This function is used by the shiny application to train
the baseline models for all the data that are stored in the var_out object
}
