% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R, R/utils.R
\name{number_of_days}
\alias{number_of_days}
\alias{number_of_days}
\title{Count the number of days}
\usage{
number_of_days(Data)

number_of_days(Data)
}
\arguments{
\item{Data}{A dataframe that contains time as column}

\item{Data}{A data frame that contains time column}
}
\value{
a numeric corresponding to the number of days
}
\description{
\code{number_of_days} This function compute the number of days for  which the data are available

\code{number_of_days} This function computes the number of days for which data
is available
}
